Feature room and user profile caching implementation wilson

+++ app/src/main/java/com/decagonhq/clads/repository/UserProfileRepositoryImpl.kt
@@ -44,29 +44,28 @@ class UserProfileRepositoryImpl(
             }
         )
 
-    override suspend fun updateUserProfile(userProfile: UserProfile): Flow<Resource<UserProfile>> =
-        networkBoundResource(
-            fetchFromLocal = {
-                database.userProfileDao().readUserProfile().map {
-                    userProfileEntityMapper.mapToDomainModel(it)
-                }
-            },
-            shouldFetchFromRemote = {
-                true
-            },
-            fetchFromRemote = {
-                delay(2000)
-                apiService.updateUserProfile(userProfileDTOMapper.mapFromDomainModel(userProfile))
-            },
-            saveToLocalDB = {
-                database.withTransaction {
-                    database.userProfileDao().deleteUserProfile()
-                    database.userProfileDao().addUserProfile(
-                        userProfileEntityMapper.mapFromDomainModel(it.payload)
-                    )
-                }
+    override suspend fun getLocalDatabaseUserProfile(): Flow<Resource<UserProfile>> {
+        return database.userProfileDao().readUserProfile().map {
+            Resource.Success(userProfileEntityMapper.mapToDomainModel(it))
+        }
+    }
+
+    override suspend fun updateUserProfile(userProfile: UserProfile) {
+        val response = safeApiCall {
+            apiService.updateUserProfile(userProfile)
+        }
+        if (response is Resource.Success) {
+            database.withTransaction {
+                database.userProfileDao().deleteUserProfile()
+                response.data?.payload?.let { userProfileEntityMapper.mapFromDomainModel(it) }
+                    ?.let {
+                        database.userProfileDao().addUserProfile(
+                            it
+                        )
+                    }
             }
-        )
+        }
+    }
 
     override suspend fun saveUserProfileToLocalDatabase() {
         val response = safeApiCall {
